etl/02_django/requirements.txt
django==3.2
flake8==4.0
django-split-settings==1.2
django-debug-toolbar==3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
uwsgi==2.0.21
gunicorn==20.1.0